Improved TopologicalSort and deprecated Loadable
