+++ src/main/java/nova/core/loader/Loadable.java
@@ -23,6 +23,7 @@
  *
  * @author Calclavia
  */
+@Deprecated
 public interface Loadable {
 	default void preInit() {
 	}
